pyproject.toml
[tool.poetry]
name = "dynamic-obs-norm"
version = "0.1.0"
description = ""
authors = ["Jet <junjet.tai@sony.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
memorial = "^0.0.8"
jj-wingman = "^0.22.1"
torch = "^2.4.1"
numpy = "^2.1.1"
gymnasium = "^1.0.0a2"
tensordict = "^0.5.0"
torchrl = "^0.5.0"
shimmy = "^2.0.0"
tqdm = "^4.66.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
